Create Numbers
This code print all the prime numbers in a 50 number interval between 0 and 400 then at the end of each line prints a count for how many prime number there are in that interval. This code finds the amount of leap years there are on a 50 year interval between 100 and 2049. This code prints the first 20 elements of the first array, then prints the first 32 elements of the first array for the text files April and June. Finally it finds the largest gap between the integers in the provided files.